remove newly added content drom serve.go and move relavent functions into server.go
